Handle Shopify Declined downgrade/upgrade
"There seems to be an issue with the billing. If I choose a plan like
the unlimited plan the billing fires and all works, but if I then
choose to downgrade the plan to the elite plan and then decline the
charge I am pushed back into the starter plan. In this can I would
assume I would remain in the unlimited plan"