Better error message message if no results found (#385)
If no videos are found during a video search, an IndexError occurs because the list of videos is empty.

The user is then shown the typical error message of Sorry! Something went wrong, we're working on it. Please try again in a few minutes.

With these changes, the code will throw an error if the list is empty and the user will be show a No videos found. error message.