docker: Fix volume overlay bug.
Because we are creating a volume with current directory => /opt/dropified
mapping, anything that is generated during the image build process is
overlayed by the contents of the volume at run time.

With this commit, we have changed the development process a bit. Now
we generate node_modules after the build process.

We are also adding scripts/setup-docker script to automate all this.