product_view.html: Add random integer to filenames. (DROP-1278) (#641)
The problem was that we were creating AWS bucket key on the basis of
filename. This was problematic when two files with same name were
uploaded, because the resulting keys for both files became same. This
commit adds a random integer to the key before uploading the file to
AWS. This reduces chance for collision.

Fixes DROP-1278