Order page cannot be viewed (DROP-1315) (#635)
* orders: Create new goto orders url.

This URL is used to get the orders from the first available store. The
problem occured when user didn't have a Shopify store. Due to this,
navigating to the orders list showed an error that store didn't exist.
However, we feel that clicking the orders list link in the sidebar
should not be affected by the existence of a store of a particular type.
So if stores of other types exist, the orders list of those stores
should be shown.

Fixes DROP-1315

* Only redirect to other platforms when that is needed

* Test fix