Fix ordering bundled products for default variants (DROP-1796) (#906)
From my research, this autocomplete section is only used for relating bundled products to the current product Default Variant, the relation is only used by get_bundle_mapping in the order's page when saving order data. Using the product id instead of variant id was the wrong id for that relation with the bundled products.